requirements.txt
discord.py 
yt-dlp 
spotipy 
pynacl 
python-dotenv